Move shared moderation code to `/antirude`
