+++ docker-compose.yml
@@ -92,6 +92,7 @@ services:
       - ./notify:/app/notify:ro
       - ./batcher:/app/batcher:ro
       - ./antispam:/app/antispam:ro
+      - ./antirude:/app/antirude:ro
       - ./async_lru_cache:/app/async_lru_cache:ro
       - ./sql:/app/sql:ro
     environment:
